Added export flag to %mf-run and also made format of flag have a '--' prefix

+++ src/metaflow_jupyter/magics.py
@@ -10,20 +10,20 @@ class MetaflowMagics(Magics):
     @cell_magic
     def mf_step(self, line, cell):
         """
-        %%mf_step <FlowName>.<StepName> [tag] [join]
+        %%mf_step <FlowName>.<StepName> [tag] [--join]
         """
         args = line.strip().split()
         if not args or "." not in args[0]:
-            raise ValueError("Explicit naming required: %%mf_step <FlowName>.<StepName> [tag] [join]")
+            raise ValueError("Explicit naming required: %%mf_step <FlowName>.<StepName> [tag] [--join]")
         
         target = args[0]
         extra_args = args[1:]
         
-        # Check if 'join' is present in any of the extra arguments
-        is_join = "join" in [a.lower() for a in extra_args]
+        # Check if '--join' is present in any of the extra arguments
+        is_join = "--join" in [a.lower() for a in extra_args]
         
         # Any argument that is NOT the word 'join' is treated as our tag
-        remaining_args = [a for a in extra_args if a.lower() != "join"]
+        remaining_args = [a for a in extra_args if a.lower() != "--join"]
         tag = remaining_args[0] if remaining_args else "default"
         
         if tag == "default":
@@ -102,14 +102,14 @@ class MetaflowMagics(Magics):
 
     @line_magic
     def mf_clear(self, line):
-        """%mf_clear <FlowName> or %mf_clear -all"""
+        """%mf_clear <FlowName> or %mf_clear --all"""
         arg = line.strip()
         if not arg:
-            print("Flow name or flag '-all' required: %mf_clear <FlowName> or %mf_clear -all")
+            print("Flow name or flag '--all' required: %mf_clear <FlowName> or %mf_clear --all")
             return
             
-        # If the arg is flag '-all', we clear registry of all flows else we only clear registry of the flow name passed
-        if "-all" in arg:
+        # If the arg is flag '--all', we clear registry of all flows else we only clear registry of the flow name passed
+        if "--all" in arg:
             registry.clear()
             print("Metaflow Registry: Cleared all flows.")
         else:
@@ -120,13 +120,16 @@ class MetaflowMagics(Magics):
     @line_magic
     def mf_run(self, line):
         """
-        %mf_run <FlowName>
+        %mf_run <FlowName> [--export]
         """
-        flow_name = line.strip()
-        if not flow_name:
-            raise ValueError("Flow name required: %mf_run <FlowName>")
+        args = line.strip().split()
+        if not args:
+            raise ValueError("Flow name required: %mf_run <FlowName> [--export]")
         
-        # conver the content of cell magics to a python script and run the script using Metaflow Runner
+        flow_name = args[0]
+        export = "--export" in [a.lower() for a in args]
+        
+        # convert the content of cell magics to a python script and run the script using Metaflow Runner
         source = registry.to_script(flow_name)
 
         filename = f"{flow_name}.py"
@@ -141,8 +144,8 @@ class MetaflowMagics(Magics):
             ErrorName = type(e).__name__
             raise Exception(f"\n{ErrorName}: {e}") from None
         finally:
-            # remove the python script after the script run completes
-            if os.path.exists(filename):
+            # remove the python script after the script run completes if export flag is not present
+            if not export and os.path.exists(filename):
                 os.remove(filename)
